+++ stores/suggestionsStore/controls.js
@@ -1,10 +1,10 @@
 export async function REQUEST_SUGGESTIONS(action) {
   const { count } = action;
   if (count) {
-    const response = await fetch(`/api/explore/people?count=${count}`);
+    const response = await fetch(`/api/suggestions/?count=${count}`);
     return await response.json();
   } else {
-    const response = await fetch(`/api/explore/people`);
+    const response = await fetch(`/api/suggestions/`);
     return await response.json();
   }
 }
