+++ stores/suggestionsStore/selectors.js
@@ -1,9 +1,5 @@
 export function getSuggestions(state, count) {
-  if (count) {
-    return state.suggestions.slice(0, count);
-  } else {
-    return state.suggestions;
-  }
+  return state.suggestions.slice(0, count);
 }
 export function getIsLoading(state) {
   return state.isLoading;
